requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.8
pandas==2.1.3
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
